pyproject.toml
[project]
name = "oneshot"
version = "1.0.0"
description = "Oneshot AI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.82.0",
    "flask>=3.1.3",
    "jinja2>=3.1.6",
    "mcp[cli]>=1.26.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "typer>=0.24.0",
    "xai-sdk>=1.7.0",
]
